requirements.txt
engineering-notation==0.6.0
pycryptodome
tk-tools==0.14.0
